refactor(ndt7): create a package for the download test
